Print centroid abundance as 64-bit
fasta_print_general() took centroid_size as int and printed it with %u,
so a centroid sequence whose abundance exceeds UINT_MAX was truncated in
the ;centroid_size= output (--centroids --centroid_sizeout). Widen the
parameter to uint64_t and print with PRIu64. The only caller passing a
non-zero value passes db_getabundance() (>= 0); all others pass 0.

+++ src/fasta.h
@@ -103,7 +103,7 @@ auto fasta_print_general(std::FILE * output_handle,
                          int clusterid,
                          char const * score_name,
                          double score,
-                         int centroid_size) -> void;
+                         uint64_t centroid_size) -> void;
 
 auto fasta_print_db(std::FILE * output_handle,
                     uint64_t seqno) -> void;
